+++ packages/common/src/types.ts
@@ -584,7 +584,7 @@ export interface BlockInfo {
     blockLastFinalized: HexString;
 
     blockHeight: bigint;
-    blockBaker: BakerId;
+    blockBaker?: BakerId;
     blockSlot: bigint;
 
     blockArriveTime: Date;
@@ -1204,6 +1204,11 @@ export interface DeployModulePayload {
     version?: number;
 
     /** Wasm module to be deployed */
+    source: Buffer;
+}
+
+export interface VersionedModuleSource {
+    version: 0 | 1;
     source: Buffer;
 }